requirements.txt
discord.py
gpt4all~=2.7.0
PyMuPDF
python-dotenv~=1.0.1
yt-dlp
pafy~=0.5.5
PyAudio~=0.2.14
gspread~=6.1.2
oauth2client~=4.1.3
pandas~=2.2.2
fuzzywuzzy~=0.18.0

discord~=2.3.2
asyncio~=3.4.3
Wave~=0.0.2
textblob~=0.18.0.post0
numpy~=1.26.4